Move ThriftRequest to tchannel.thrift.rw
